+++ upcloud/problem.go
@@ -2,6 +2,7 @@ package upcloud
 
 import (
 	"fmt"
+	"net/url"
 	"strings"
 )
 
@@ -14,10 +15,17 @@ type Problem struct {
 	// InvalidParams if set, is a list of ProblemInvalidParam describing a specific part(s) of the request
 	// that caused the problem
 	InvalidParams []ProblemInvalidParam `json:"invalid_params,omitempty"`
-	// CorrelationID is an unique string that identifies the request that caused the problem
+	// CorrelationID is a unique string that identifies the request that caused the problem
+	// Please note that it is not always available
 	CorrelationID string `json:"correlation_id,omitempty"`
 	// HTTP Status code
 	Status int `json:"status"`
+}
+
+// ProblemInvalidParam is a type describing extra information in the Problem type's InvalidParams field.
+type ProblemInvalidParam struct {
+	Name   string `json:"name"`
+	Reason string `json:"reason"`
 }
 
 func (p *Problem) Error() string {
@@ -34,8 +42,15 @@ func (p *Problem) Error() string {
 	return sb.String()
 }
 
-// ProblemInvalidParam is a type describing extra information in the Problem type's InvalidParams field.
-type ProblemInvalidParam struct {
-	Name   string `json:"name"`
-	Reason string `json:"reason"`
+// ErrorCode returns a short string that identifies the error; it should be used for programmatic comparisons
+func (p *Problem) ErrorCode() string {
+	// First check if the type is a URL.
+	// If it is - we need to extract meaningful fragment from it for comparison purposes
+	// If it isn't - we can just return the value of `Type` field
+	parsedURL, err := url.Parse(p.Type)
+	if err != nil || parsedURL.Scheme == "" || parsedURL.Host == "" {
+		return p.Type
+	}
+
+	return strings.Replace(parsedURL.Fragment, "ERROR_", "", 1)
 }