+++ CHANGELOG.md
@@ -5,6 +5,12 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Changed
+- errors: all service method now return `Problem` type in case of errors (*BREAKING CHANGE*)
+
+### Removed
+- errors: `Error` type was removed (*BREAKING CHANGE*)
+
 ## [5.4.0]
 
 ### Added